ConnectionMultiplexer: split and cleanup
Mostly prep for more changes:
- Splitting functionality to partials
- Ordering methods
- Cleanup instance ordering
