+++ src/StackExchange.Redis/ConnectionMultiplexer.FeatureFlags.cs
@@ -1,0 +1,56 @@
+﻿using System;
+using System.ComponentModel;
+using System.Threading;
+
+namespace StackExchange.Redis;
+
+public partial class ConnectionMultiplexer
+{
+    private static FeatureFlags s_featureFlags;
+
+    [Flags]
+    private enum FeatureFlags
+    {
+        None,
+        PreventThreadTheft = 1,
+    }
+
+    private static void SetAutodetectFeatureFlags()
+    {
+        bool value = false;
+        try
+        {   // attempt to detect a known problem scenario
+            value = SynchronizationContext.Current?.GetType()?.Name
+                == "LegacyAspNetSynchronizationContext";
+        }
+        catch { }
+        SetFeatureFlag(nameof(FeatureFlags.PreventThreadTheft), value);
+    }
+
+    /// <summary>
+    /// Enables or disables a feature flag.
+    /// This should only be used under support guidance, and should not be rapidly toggled.
+    /// </summary>
+    [EditorBrowsable(EditorBrowsableState.Never)]
+    [Browsable(false)]
+    public static void SetFeatureFlag(string flag, bool enabled)
+    {
+        if (Enum.TryParse<FeatureFlags>(flag, true, out var flags))
+        {
+            if (enabled) s_featureFlags |= flags;
+            else s_featureFlags &= ~flags;
+        }
+    }
+
+    /// <summary>
+    /// Returns the state of a feature flag.
+    /// This should only be used under support guidance.
+    /// </summary>
+    [EditorBrowsable(EditorBrowsableState.Never)]
+    [Browsable(false)]
+    public static bool GetFeatureFlag(string flag)
+        => Enum.TryParse<FeatureFlags>(flag, true, out var flags)
+        && (s_featureFlags & flags) == flags;
+
+    internal static bool PreventThreadTheft => (s_featureFlags & FeatureFlags.PreventThreadTheft) != 0;
+}
