+++ src/StackExchange.Redis/ConnectionMultiplexer.Threading.cs
@@ -2,49 +2,48 @@
 using System.Threading;
 using Pipelines.Sockets.Unofficial;
 
-namespace StackExchange.Redis
-{
-    public partial class ConnectionMultiplexer
-    {
-        private static readonly WaitCallback s_CompleteAsWorker = s => ((ICompletable)s).TryComplete(true);
-        internal static void CompleteAsWorker(ICompletable completable)
-        {
-            if (completable != null)
-            {
-                ThreadPool.QueueUserWorkItem(s_CompleteAsWorker, completable);
-            }
-        }
+namespace StackExchange.Redis;
 
-        internal static bool TryCompleteHandler<T>(EventHandler<T> handler, object sender, T args, bool isAsync) where T : EventArgs, ICompletable
+public partial class ConnectionMultiplexer
+{
+    private static readonly WaitCallback s_CompleteAsWorker = s => ((ICompletable)s).TryComplete(true);
+    internal static void CompleteAsWorker(ICompletable completable)
+    {
+        if (completable != null)
         {
-            if (handler == null) return true;
-            if (isAsync)
+            ThreadPool.QueueUserWorkItem(s_CompleteAsWorker, completable);
+        }
+    }
+
+    internal static bool TryCompleteHandler<T>(EventHandler<T> handler, object sender, T args, bool isAsync) where T : EventArgs, ICompletable
+    {
+        if (handler == null) return true;
+        if (isAsync)
+        {
+            if (handler.IsSingle())
             {
-                if (handler.IsSingle())
+                try
                 {
-                    try
-                    {
-                        handler(sender, args);
-                    }
-                    catch { }
+                    handler(sender, args);
                 }
-                else
-                {
-                    foreach (EventHandler<T> sub in handler.AsEnumerable())
-                    {
-                        try
-                        {
-                            sub(sender, args);
-                        }
-                        catch { }
-                    }
-                }
-                return true;
+                catch { }
             }
             else
             {
-                return false;
+                foreach (EventHandler<T> sub in handler.AsEnumerable())
+                {
+                    try
+                    {
+                        sub(sender, args);
+                    }
+                    catch { }
+                }
             }
+            return true;
+        }
+        else
+        {
+            return false;
         }
     }
 }
