+++ src/StackExchange.Redis/ConnectionMultiplexer.cs
@@ -1,9 +1,7 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.IO;
-using System.IO.Compression;
 using System.Linq;
 using System.Net;
 using System.Runtime.CompilerServices;
@@ -12,7 +10,6 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Pipelines.Sockets.Unofficial;
-using StackExchange.Redis.Maintenance;
 using StackExchange.Redis.Profiling;
 
 namespace StackExchange.Redis
@@ -24,117 +21,61 @@ namespace StackExchange.Redis
     /// <remarks>https://stackexchange.github.io/StackExchange.Redis/PipelinesMultiplexers</remarks>
     public sealed partial class ConnectionMultiplexer : IInternalConnectionMultiplexer // implies : IConnectionMultiplexer and : IDisposable
     {
-        [Flags]
-        private enum FeatureFlags
-        {
-            None,
-            PreventThreadTheft = 1,
-        }
+        internal const int MillisecondsPerHeartbeat = 1000;
 
-        private static FeatureFlags s_featureFlags;
-
-        /// <summary>
-        /// Enables or disables a feature flag.
-        /// This should only be used under support guidance, and should not be rapidly toggled.
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        [Browsable(false)]
-        public static void SetFeatureFlag(string flag, bool enabled)
-        {
-            if (Enum.TryParse<FeatureFlags>(flag, true, out var flags))
-            {
-                if (enabled) s_featureFlags |= flags;
-                else s_featureFlags &= ~flags;
-            }
-        }
-
-        static ConnectionMultiplexer()
-        {
-            bool value = false;
-            try
-            {   // attempt to detect a known problem scenario
-                value = SynchronizationContext.Current?.GetType()?.Name
-                    == "LegacyAspNetSynchronizationContext";
-            }
-            catch { }
-            SetFeatureFlag(nameof(FeatureFlags.PreventThreadTheft), value);
-        }
-
-        /// <summary>
-        /// Returns the state of a feature flag.
-        /// This should only be used under support guidance.
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        [Browsable(false)]
-        public static bool GetFeatureFlag(string flag)
-            => Enum.TryParse<FeatureFlags>(flag, true, out var flags)
-            && (s_featureFlags & flags) == flags;
-
-        internal static bool PreventThreadTheft => (s_featureFlags & FeatureFlags.PreventThreadTheft) != 0;
-
-#if DEBUG
-        private static int _collectedWithoutDispose;
-        internal static int CollectedWithoutDispose => Thread.VolatileRead(ref _collectedWithoutDispose);
-        /// <summary>
-        /// Invoked by the garbage collector.
-        /// </summary>
-        ~ConnectionMultiplexer()
-        {
-            Interlocked.Increment(ref _collectedWithoutDispose);
-        }
-#endif
-
-        bool IInternalConnectionMultiplexer.AllowConnect
-        {
-            get => AllowConnect;
-            set => AllowConnect = value;
-        }
-
-        bool IInternalConnectionMultiplexer.IgnoreConnect
-        {
-            get => IgnoreConnect;
-            set => IgnoreConnect = value;
-        }
-
-        /// <summary>
-        /// For debugging: when not enabled, servers cannot connect.
-        /// </summary>
-        internal volatile bool AllowConnect = true;
-
-        /// <summary>
-        /// For debugging: when not enabled, end-connect is silently ignored (to simulate a long-running connect).
-        /// </summary>
-        internal volatile bool IgnoreConnect;
+        // This gets accessed for every received event; let's make sure we can process it "raw"
+        internal readonly byte[] ConfigurationChangedChannel;
+        // Unique identifier used when tracing
+        internal readonly byte[] UniqueId = Guid.NewGuid().ToByteArray();
 
         /// <summary>
         /// Tracks overall connection multiplexer counts.
         /// </summary>
         internal int _connectAttemptCount = 0, _connectCompletedCount = 0, _connectionCloseCount = 0;
+        private long syncTimeouts, fireAndForgets, asyncTimeouts;
+        private string failureMessage, activeConfigCause;
+        private IDisposable pulse;
+
+        private readonly Hashtable servers = new Hashtable();
+        private volatile ServerSnapshot _serverSnapshot = ServerSnapshot.Empty;
+
+        private volatile bool _isDisposed;
+        internal bool IsDisposed => _isDisposed;
+
+        internal CommandMap CommandMap { get; }
+        internal ConfigurationOptions RawConfig { get; }
+        internal ServerSelectionStrategy ServerSelectionStrategy { get; }
+        internal Exception LastException { get; set; }
+
+        private int _activeHeartbeatErrors, lastHeartbeatTicks;
+        internal long LastHeartbeatSecondsAgo =>
+            pulse is null
+            ? -1
+            : unchecked(Environment.TickCount - Thread.VolatileRead(ref lastHeartbeatTicks)) / 1000;
+
+        private static int lastGlobalHeartbeatTicks = Environment.TickCount;
+        internal static long LastGlobalHeartbeatSecondsAgo =>
+            unchecked(Environment.TickCount - Thread.VolatileRead(ref lastGlobalHeartbeatTicks)) / 1000;
 
         /// <summary>
-        /// No longer used.
+        /// Should exceptions include identifiable details? (key names, additional .Data annotations)
         /// </summary>
-        [Obsolete("No longer used, will be removed in 3.0.")]
-        public static TaskFactory Factory
-        {
-            get => Task.Factory;
-            set { }
-        }
+        public bool IncludeDetailInExceptions { get; set; }
 
         /// <summary>
-        /// Get summary statistics associated with all servers in this multiplexer.
+        /// Should exceptions include performance counter details? (CPU usage, etc - note that this can be problematic on some platforms)
         /// </summary>
-        public ServerCounters GetCounters()
-        {
-            var snapshot = GetServerSnapshot();
+        public bool IncludePerformanceCountersInExceptions { get; set; }
 
-            var counters = new ServerCounters(null);
-            for (int i = 0; i < snapshot.Length; i++)
-            {
-                counters.Add(snapshot[i].GetCounters());
-            }
-            return counters;
-        }
+        /// <summary>
+        /// Gets the synchronous timeout associated with the connections.
+        /// </summary>
+        public int TimeoutMilliseconds { get; }
+
+        /// <summary>
+        /// Gets the asynchronous timeout associated with the connections.
+        /// </summary>
+        internal int AsyncTimeoutMilliseconds { get; }
 
         /// <summary>
         /// Gets the client-name that will be used on all new connections.
@@ -149,195 +90,128 @@ namespace StackExchange.Redis
         /// </summary>
         public string Configuration => RawConfig.ToString();
 
-        internal void OnConnectionFailed(EndPoint endpoint, ConnectionType connectionType, ConnectionFailureType failureType, Exception exception, bool reconfigure, string physicalName)
+        /// <summary>
+        /// Indicates whether any servers are connected.
+        /// </summary>
+        public bool IsConnected
         {
-            if (_isDisposed) return;
-            var handler = ConnectionFailed;
-            if (handler != null)
+            get
             {
-                CompleteAsWorker(new ConnectionFailedEventArgs(handler, this, endpoint, connectionType, failureType, exception, physicalName));
-            }
-            if (reconfigure)
-            {
-                ReconfigureIfNeeded(endpoint, false, "connection failed");
+                var tmp = GetServerSnapshot();
+                for (int i = 0; i < tmp.Length; i++)
+                    if (tmp[i].IsConnected) return true;
+                return false;
             }
         }
-
-        internal void OnInternalError(Exception exception, EndPoint endpoint = null, ConnectionType connectionType = ConnectionType.None, [CallerMemberName] string origin = null)
-        {
-            try
-            {
-                if (_isDisposed) return;
-                Trace("Internal error: " + origin + ", " + exception == null ? "unknown" : exception.Message);
-                var handler = InternalError;
-                if (handler != null)
-                {
-                    CompleteAsWorker(new InternalErrorEventArgs(handler, this, endpoint, connectionType, exception, origin));
-                }
-            }
-            catch
-            {
-                // Our internal error event failed...whatcha gonna do, exactly?
-            }
-        }
-
-        internal void OnConnectionRestored(EndPoint endpoint, ConnectionType connectionType, string physicalName)
-        {
-            if (_isDisposed) return;
-            var handler = ConnectionRestored;
-            if (handler != null)
-            {
-                CompleteAsWorker(new ConnectionFailedEventArgs(handler, this, endpoint, connectionType, ConnectionFailureType.None, null, physicalName));
-            }
-            ReconfigureIfNeeded(endpoint, false, "connection restored");
-        }
-
-        private void OnEndpointChanged(EndPoint endpoint, EventHandler<EndPointEventArgs> handler)
-        {
-            if (_isDisposed) return;
-            if (handler != null)
-            {
-                CompleteAsWorker(new EndPointEventArgs(handler, this, endpoint));
-            }
-        }
-
-        internal void OnConfigurationChanged(EndPoint endpoint) => OnEndpointChanged(endpoint, ConfigurationChanged);
-        internal void OnConfigurationChangedBroadcast(EndPoint endpoint) => OnEndpointChanged(endpoint, ConfigurationChangedBroadcast);
 
         /// <summary>
-        /// Raised when a server replied with an error message.
+        /// Indicates whether any servers are currently trying to connect.
         /// </summary>
-        public event EventHandler<RedisErrorEventArgs> ErrorMessage;
-        internal void OnErrorMessage(EndPoint endpoint, string message)
+        public bool IsConnecting
         {
-            if (_isDisposed) return;
-            var handler = ErrorMessage;
-            if (handler != null)
+            get
             {
-                CompleteAsWorker(new RedisErrorEventArgs(handler, this, endpoint, message));
+                var tmp = GetServerSnapshot();
+                for (int i = 0; i < tmp.Length; i++)
+                    if (tmp[i].IsConnecting) return true;
+                return false;
             }
         }
 
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2202:Do not dispose objects multiple times")]
-        private static void Write<T>(ZipArchive zip, string name, Task task, Action<T, StreamWriter> callback)
+        static ConnectionMultiplexer()
         {
-            var entry = zip.CreateEntry(name, CompressionLevel.Optimal);
-            using (var stream = entry.Open())
-            using (var writer = new StreamWriter(stream))
-            {
-                TaskStatus status = task.Status;
-                switch (status)
-                {
-                    case TaskStatus.RanToCompletion:
-                        T val = ((Task<T>)task).Result;
-                        callback(val, writer);
-                        break;
-                    case TaskStatus.Faulted:
-                        writer.WriteLine(string.Join(", ", task.Exception.InnerExceptions.Select(x => x.Message)));
-                        break;
-                    default:
-                        writer.WriteLine(status.ToString());
-                        break;
-                }
-            }
+            SetAutodetectFeatureFlags();
         }
+
+        private ConnectionMultiplexer(ConfigurationOptions configuration)
+        {
+            IncludeDetailInExceptions = true;
+            IncludePerformanceCountersInExceptions = false;
+
+            RawConfig = configuration ?? throw new ArgumentNullException(nameof(configuration));
+
+            var map = CommandMap = configuration.CommandMap;
+            if (!string.IsNullOrWhiteSpace(configuration.Password)) map.AssertAvailable(RedisCommand.AUTH);
+
+            if (!map.IsAvailable(RedisCommand.ECHO) && !map.IsAvailable(RedisCommand.PING) && !map.IsAvailable(RedisCommand.TIME))
+            {
+                // I mean really, give me a CHANCE! I need *something* to check the server is available to me...
+                // see also: SendTracer (matching logic)
+                map.AssertAvailable(RedisCommand.EXISTS);
+            }
+
+            TimeoutMilliseconds = configuration.SyncTimeout;
+            AsyncTimeoutMilliseconds = configuration.AsyncTimeout;
+
+            OnCreateReaderWriter(configuration);
+            ServerSelectionStrategy = new ServerSelectionStrategy(this);
+
+            var configChannel = configuration.ConfigurationChannel;
+            if (!string.IsNullOrWhiteSpace(configChannel))
+            {
+                ConfigurationChangedChannel = Encoding.UTF8.GetBytes(configChannel);
+            }
+            lastHeartbeatTicks = Environment.TickCount;
+        }
+
+        private static ConnectionMultiplexer CreateMultiplexer(ConfigurationOptions configuration, LogProxy log, out EventHandler<ConnectionFailedEventArgs> connectHandler)
+        {
+            var muxer = new ConnectionMultiplexer(configuration);
+            connectHandler = null;
+            if (log is not null)
+            {
+                // Create a detachable event-handler to log detailed errors if something happens during connect/handshake
+                connectHandler = (_, a) =>
+                {
+                    try
+                    {
+                        lock (log.SyncLock) // Keep the outer and any inner errors contiguous
+                        {
+                            var ex = a.Exception;
+                            log?.WriteLine($"Connection failed: {Format.ToString(a.EndPoint)} ({a.ConnectionType}, {a.FailureType}): {ex?.Message ?? "(unknown)"}");
+                            while ((ex = ex.InnerException) != null)
+                            {
+                                log?.WriteLine($"> {ex.Message}");
+                            }
+                        }
+                    }
+                    catch { }
+                };
+                muxer.ConnectionFailed += connectHandler;
+            }
+            return muxer;
+        }
+
         /// <summary>
-        /// Write the configuration of all servers to an output stream.
+        /// Get summary statistics associated with all servers in this multiplexer.
         /// </summary>
-        /// <param name="destination">The destination stream to write the export to.</param>
-        /// <param name="options">The options to use for this export.</param>
-        public void ExportConfiguration(Stream destination, ExportOptions options = ExportOptions.All)
+        public ServerCounters GetCounters()
         {
-            if (destination == null) throw new ArgumentNullException(nameof(destination));
-
-            // What is possible, given the command map?
-            ExportOptions mask = 0;
-            if (CommandMap.IsAvailable(RedisCommand.INFO)) mask |= ExportOptions.Info;
-            if (CommandMap.IsAvailable(RedisCommand.CONFIG)) mask |= ExportOptions.Config;
-            if (CommandMap.IsAvailable(RedisCommand.CLIENT)) mask |= ExportOptions.Client;
-            if (CommandMap.IsAvailable(RedisCommand.CLUSTER)) mask |= ExportOptions.Cluster;
-            options &= mask;
-
-            using (var zip = new ZipArchive(destination, ZipArchiveMode.Create, true))
+            var counters = new ServerCounters(null);
+            var snapshot = GetServerSnapshot();
+            for (int i = 0; i < snapshot.Length; i++)
             {
-                var arr = GetServerSnapshot();
-                foreach (var server in arr)
-                {
-                    const CommandFlags flags = CommandFlags.None;
-                    if (!server.IsConnected) continue;
-                    var api = GetServer(server.EndPoint);
-
-                    List<Task> tasks = new List<Task>();
-                    if ((options & ExportOptions.Info) != 0)
-                    {
-                        tasks.Add(api.InfoRawAsync(flags: flags));
-                    }
-                    if ((options & ExportOptions.Config) != 0)
-                    {
-                        tasks.Add(api.ConfigGetAsync(flags: flags));
-                    }
-                    if ((options & ExportOptions.Client) != 0)
-                    {
-                        tasks.Add(api.ClientListAsync(flags: flags));
-                    }
-                    if ((options & ExportOptions.Cluster) != 0)
-                    {
-                        tasks.Add(api.ClusterNodesRawAsync(flags: flags));
-                    }
-
-                    WaitAllIgnoreErrors(tasks.ToArray());
-
-                    int index = 0;
-                    var prefix = Format.ToString(server.EndPoint);
-                    if ((options & ExportOptions.Info) != 0)
-                    {
-                        Write<string>(zip, prefix + "/info.txt", tasks[index++], WriteNormalizingLineEndings);
-                    }
-                    if ((options & ExportOptions.Config) != 0)
-                    {
-                        Write<KeyValuePair<string, string>[]>(zip, prefix + "/config.txt", tasks[index++], (settings, writer) =>
-                        {
-                            foreach (var setting in settings)
-                            {
-                                writer.WriteLine("{0}={1}", setting.Key, setting.Value);
-                            }
-                        });
-                    }
-                    if ((options & ExportOptions.Client) != 0)
-                    {
-                        Write<ClientInfo[]>(zip, prefix + "/clients.txt", tasks[index++], (clients, writer) =>
-                        {
-                            if (clients == null)
-                            {
-                                writer.WriteLine(NoContent);
-                            }
-                            else
-                            {
-                                foreach (var client in clients)
-                                {
-                                    writer.WriteLine(client.Raw);
-                                }
-                            }
-                        });
-                    }
-                    if ((options & ExportOptions.Cluster) != 0)
-                    {
-                        Write<string>(zip, prefix + "/nodes.txt", tasks[index++], WriteNormalizingLineEndings);
-                    }
-                }
+                counters.Add(snapshot[i].GetCounters());
             }
+            return counters;
         }
 
         internal async Task MakePrimaryAsync(ServerEndPoint server, ReplicationChangeOptions options, LogProxy log)
         {
+            _ = server ?? throw new ArgumentNullException(nameof(server));
+
             var cmd = server.GetFeatures().ReplicaCommands ? RedisCommand.REPLICAOF : RedisCommand.SLAVEOF;
             CommandMap.AssertAvailable(cmd);
 
-            if (!RawConfig.AllowAdmin) throw ExceptionFactory.AdminModeNotEnabled(IncludeDetailInExceptions, cmd, null, server);
-
-            if (server == null) throw new ArgumentNullException(nameof(server));
+            if (!RawConfig.AllowAdmin)
+            {
+                throw ExceptionFactory.AdminModeNotEnabled(IncludeDetailInExceptions, cmd, null, server);
+            }
             var srv = new RedisServer(this, server, null);
-            if (!srv.IsConnected) throw ExceptionFactory.NoConnectionAvailable(this, null, server, GetServerSnapshot(), command: cmd);
+            if (!srv.IsConnected)
+            {
+                throw ExceptionFactory.NoConnectionAvailable(this, null, server, GetServerSnapshot(), command: cmd);
+            }
 
             const CommandFlags flags = CommandFlags.NoRedirect;
             Message msg;
@@ -358,7 +232,8 @@ namespace StackExchange.Redis
 
             RedisKey tieBreakerKey = default(RedisKey);
             // try and write this everywhere; don't worry if some folks reject our advances
-            if ((options & ReplicationChangeOptions.SetTiebreaker) != 0 && !string.IsNullOrWhiteSpace(RawConfig.TieBreaker)
+            if (options.HasFlag(ReplicationChangeOptions.SetTiebreaker)
+                && !string.IsNullOrWhiteSpace(RawConfig.TieBreaker)
                 && CommandMap.IsAvailable(RedisCommand.SET))
             {
                 tieBreakerKey = RawConfig.TieBreaker;
@@ -414,7 +289,8 @@ namespace StackExchange.Redis
             // This eliminates the race of pub/sub *then* re-slaving happening, since a method both precedes and follows.
             async Task BroadcastAsync(ServerEndPoint[] serverNodes)
             {
-                if ((options & ReplicationChangeOptions.Broadcast) != 0 && ConfigurationChangedChannel != null
+                if (options.HasFlag(ReplicationChangeOptions.Broadcast)
+                    && ConfigurationChangedChannel != null
                     && CommandMap.IsAvailable(RedisCommand.PUBLISH))
                 {
                     RedisValue channel = ConfigurationChangedChannel;
@@ -431,7 +307,7 @@ namespace StackExchange.Redis
             // Send a message before it happens - because afterwards a new replica may be unresponsive
             await BroadcastAsync(nodes);
 
-            if ((options & ReplicationChangeOptions.ReplicateToOtherEndpoints) != 0)
+            if (options.HasFlag(ReplicationChangeOptions.ReplicateToOtherEndpoints))
             {
                 foreach (var node in nodes)
                 {
@@ -465,89 +341,23 @@ namespace StackExchange.Redis
         internal void CheckMessage(Message message)
         {
             if (!RawConfig.AllowAdmin && message.IsAdmin)
+            {
                 throw ExceptionFactory.AdminModeNotEnabled(IncludeDetailInExceptions, message.Command, message, null);
-            if (message.Command != RedisCommand.UNKNOWN) CommandMap.AssertAvailable(message.Command);
+            }
+            if (message.Command != RedisCommand.UNKNOWN)
+            {
+                CommandMap.AssertAvailable(message.Command);
+            }
 
             // using >= here because we will be adding 1 for the command itself (which is an argument for the purposes of the multi-bulk protocol)
-            if (message.ArgCount >= PhysicalConnection.REDIS_MAX_ARGS) throw ExceptionFactory.TooManyArgs(message.CommandAndKey, message.ArgCount);
-        }
-        private const string NoContent = "(no content)";
-        private static void WriteNormalizingLineEndings(string source, StreamWriter writer)
-        {
-            if (source == null)
+            if (message.ArgCount >= PhysicalConnection.REDIS_MAX_ARGS)
             {
-                writer.WriteLine(NoContent);
-            }
-            else
-            {
-                using (var reader = new StringReader(source))
-                {
-                    string line;
-                    while ((line = reader.ReadLine()) != null)
-                        writer.WriteLine(line); // normalize line endings
-                }
+                throw ExceptionFactory.TooManyArgs(message.CommandAndKey, message.ArgCount);
             }
         }
 
-        /// <summary>
-        /// Raised whenever a physical connection fails.
-        /// </summary>
-        public event EventHandler<ConnectionFailedEventArgs> ConnectionFailed;
-
-        /// <summary>
-        /// Raised whenever an internal error occurs (this is primarily for debugging).
-        /// </summary>
-        public event EventHandler<InternalErrorEventArgs> InternalError;
-
-        /// <summary>
-        /// Raised whenever a physical connection is established.
-        /// </summary>
-        public event EventHandler<ConnectionFailedEventArgs> ConnectionRestored;
-
-        /// <summary>
-        /// Raised when configuration changes are detected.
-        /// </summary>
-        public event EventHandler<EndPointEventArgs> ConfigurationChanged;
-
-        /// <summary>
-        /// Raised when nodes are explicitly requested to reconfigure via broadcast.
-        /// This usually means primary/replica changes.
-        /// </summary>
-        public event EventHandler<EndPointEventArgs> ConfigurationChangedBroadcast;
-
-        /// <summary>
-        /// Raised when server indicates a maintenance event is going to happen.
-        /// </summary>
-        public event EventHandler<ServerMaintenanceEvent> ServerMaintenanceEvent;
-
-        /// <summary>
-        /// Gets the synchronous timeout associated with the connections.
-        /// </summary>
-        public int TimeoutMilliseconds { get; }
-
-        /// <summary>
-        /// Gets the asynchronous timeout associated with the connections.
-        /// </summary>
-        internal int AsyncTimeoutMilliseconds { get; }
-
-        /// <summary>
-        /// Gets all endpoints defined on the multiplexer.
-        /// </summary>
-        /// <param name="configuredOnly">Whether to get only the endpoints specified explicitly in the config.</param>
-        public EndPoint[] GetEndPoints(bool configuredOnly = false)
-        {
-            if (configuredOnly) return RawConfig.EndPoints.ToArray();
-
-            return _serverSnapshot.GetEndPoints();
-        }
-
-        internal void InvokeServerMaintenanceEvent(ServerMaintenanceEvent e)
-            => ServerMaintenanceEvent?.Invoke(this, e);
-
-        internal bool TryResend(int hashSlot, Message message, EndPoint endpoint, bool isMoved)
-        {
-            return ServerSelectionStrategy.TryResend(hashSlot, message, endpoint, isMoved);
-        }
+        internal bool TryResend(int hashSlot, Message message, EndPoint endpoint, bool isMoved) =>
+            ServerSelectionStrategy.TryResend(hashSlot, message, endpoint, isMoved);
 
         /// <summary>
         /// Wait for a given asynchronous operation to complete (or timeout).
@@ -555,10 +365,13 @@ namespace StackExchange.Redis
         /// <param name="task">The task to wait on.</param>
         public void Wait(Task task)
         {
-            if (task == null) throw new ArgumentNullException(nameof(task));
+            _ = task ?? throw new ArgumentNullException(nameof(task));
             try
             {
-                if (!task.Wait(TimeoutMilliseconds)) throw new TimeoutException();
+                if (!task.Wait(TimeoutMilliseconds))
+                {
+                    throw new TimeoutException();
+                }
             }
             catch (AggregateException aex) when (IsSingle(aex))
             {
@@ -573,10 +386,13 @@ namespace StackExchange.Redis
         /// <param name="task">The task to wait on.</param>
         public T Wait<T>(Task<T> task)
         {
-            if (task == null) throw new ArgumentNullException(nameof(task));
+            _ = task ?? throw new ArgumentNullException(nameof(task));
             try
             {
-                if (!task.Wait(TimeoutMilliseconds)) throw new TimeoutException();
+                if (!task.Wait(TimeoutMilliseconds))
+                {
+                    throw new TimeoutException();
+                }
             }
             catch (AggregateException aex) when (IsSingle(aex))
             {
@@ -587,8 +403,14 @@ namespace StackExchange.Redis
 
         private static bool IsSingle(AggregateException aex)
         {
-            try { return aex != null && aex.InnerExceptions.Count == 1; }
-            catch { return false; }
+            try
+            {
+                return aex?.InnerExceptions.Count == 1;
+            }
+            catch
+            {
+                return false;
+            }
         }
 
         /// <summary>
@@ -597,22 +419,23 @@ namespace StackExchange.Redis
         /// <param name="tasks">The tasks to wait on.</param>
         public void WaitAll(params Task[] tasks)
         {
-            if (tasks == null) throw new ArgumentNullException(nameof(tasks));
+            _ = tasks ?? throw new ArgumentNullException(nameof(tasks));
             if (tasks.Length == 0) return;
-            if (!Task.WaitAll(tasks, TimeoutMilliseconds)) throw new TimeoutException();
+            if (!Task.WaitAll(tasks, TimeoutMilliseconds))
+            {
+                throw new TimeoutException();
+            }
         }
 
-        private bool WaitAllIgnoreErrors(Task[] tasks) => WaitAllIgnoreErrors(tasks, TimeoutMilliseconds);
-
-        private static bool WaitAllIgnoreErrors(Task[] tasks, int timeout)
+        private bool WaitAllIgnoreErrors(Task[] tasks)
         {
-            if (tasks == null) throw new ArgumentNullException(nameof(tasks));
+            _ = tasks ?? throw new ArgumentNullException(nameof(tasks));
             if (tasks.Length == 0) return true;
             var watch = ValueStopwatch.StartNew();
             try
             {
                 // If no error, great
-                if (Task.WaitAll(tasks, timeout)) return true;
+                if (Task.WaitAll(tasks, TimeoutMilliseconds)) return true;
             }
             catch
             { }
@@ -622,7 +445,7 @@ namespace StackExchange.Redis
                 var task = tasks[i];
                 if (!task.IsCanceled && !task.IsCompleted && !task.IsFaulted)
                 {
-                    var remaining = timeout - watch.ElapsedMilliseconds;
+                    var remaining = TimeoutMilliseconds - watch.ElapsedMilliseconds;
                     if (remaining <= 0) return false;
                     try
                     {
@@ -635,50 +458,35 @@ namespace StackExchange.Redis
             return false;
         }
 
-        internal bool AuthSuspect { get; private set; }
-        internal void SetAuthSuspect() => AuthSuspect = true;
-
-        private static void LogWithThreadPoolStats(LogProxy log, string message, out int busyWorkerCount)
-        {
-            busyWorkerCount = 0;
-            if (log != null)
-            {
-                var sb = new StringBuilder();
-                sb.Append(message);
-                busyWorkerCount = PerfCounterHelper.GetThreadPoolStats(out string iocp, out string worker, out string workItems);
-                sb.Append(", IOCP: ").Append(iocp).Append(", WORKER: ").Append(worker);
-                if (workItems != null)
-                {
-                    sb.Append(", POOL: ").Append(workItems);
-                }
-                log?.WriteLine(sb.ToString());
-            }
-        }
-
-        private static bool AllComplete(Task[] tasks)
-        {
-            for (int i = 0; i < tasks.Length; i++)
-            {
-                var task = tasks[i];
-                if (!task.IsCanceled && !task.IsCompleted && !task.IsFaulted)
-                    return false;
-            }
-            return true;
-        }
-
         private static async Task<bool> WaitAllIgnoreErrorsAsync(string name, Task[] tasks, int timeoutMilliseconds, LogProxy log, [CallerMemberName] string caller = null, [CallerLineNumber] int callerLineNumber = 0)
         {
-            if (tasks == null) throw new ArgumentNullException(nameof(tasks));
+            _ = tasks ?? throw new ArgumentNullException(nameof(tasks));
             if (tasks.Length == 0)
             {
                 log?.WriteLine("No tasks to await");
                 return true;
             }
-
             if (AllComplete(tasks))
             {
                 log?.WriteLine("All tasks are already complete");
                 return true;
+            }
+
+            static void LogWithThreadPoolStats(LogProxy log, string message, out int busyWorkerCount)
+            {
+                busyWorkerCount = 0;
+                if (log != null)
+                {
+                    var sb = new StringBuilder();
+                    sb.Append(message);
+                    busyWorkerCount = PerfCounterHelper.GetThreadPoolStats(out string iocp, out string worker, out string workItems);
+                    sb.Append(", IOCP: ").Append(iocp).Append(", WORKER: ").Append(worker);
+                    if (workItems != null)
+                    {
+                        sb.Append(", POOL: ").Append(workItems);
+                    }
+                    log?.WriteLine(sb.ToString());
+                }
             }
 
             var watch = ValueStopwatch.StartNew();
@@ -726,67 +534,19 @@ namespace StackExchange.Redis
             return false;
         }
 
-        /// <summary>
-        /// Raised when a hash-slot has been relocated.
-        /// </summary>
-        public event EventHandler<HashSlotMovedEventArgs> HashSlotMoved;
-
-        internal void OnHashSlotMoved(int hashSlot, EndPoint old, EndPoint @new)
+        private static bool AllComplete(Task[] tasks)
         {
-            var handler = HashSlotMoved;
-            if (handler != null)
+            for (int i = 0; i < tasks.Length; i++)
             {
-                CompleteAsWorker(new HashSlotMovedEventArgs(handler, this, hashSlot, old, @new));
+                var task = tasks[i];
+                if (!task.IsCanceled && !task.IsCompleted && !task.IsFaulted)
+                    return false;
             }
+            return true;
         }
 
-        /// <summary>
-        /// Compute the hash-slot of a specified key.
-        /// </summary>
-        /// <param name="key">The key to get a hash slot ID for.</param>
-        public int HashSlot(RedisKey key) => ServerSelectionStrategy.HashSlot(key);
-
-        internal ServerEndPoint AnyServer(ServerType serverType, uint startOffset, RedisCommand command, CommandFlags flags, bool allowDisconnected)
-        {
-            var tmp = GetServerSnapshot();
-            int len = tmp.Length;
-            ServerEndPoint fallback = null;
-            for (int i = 0; i < len; i++)
-            {
-                var server = tmp[(int)(((uint)i + startOffset) % len)];
-                if (server != null && server.ServerType == serverType && server.IsSelectable(command, allowDisconnected))
-                {
-                    if (server.IsReplica)
-                    {
-                        switch (flags)
-                        {
-                            case CommandFlags.DemandReplica:
-                            case CommandFlags.PreferReplica:
-                                return server;
-                            case CommandFlags.PreferMaster:
-                                fallback = server;
-                                break;
-                        }
-                    }
-                    else
-                    {
-                        switch (flags)
-                        {
-                            case CommandFlags.DemandMaster:
-                            case CommandFlags.PreferMaster:
-                                return server;
-                            case CommandFlags.PreferReplica:
-                                fallback = server;
-                                break;
-                        }
-                    }
-                }
-            }
-            return fallback;
-        }
-
-        private volatile bool _isDisposed;
-        internal bool IsDisposed => _isDisposed;
+        internal bool AuthSuspect { get; private set; }
+        internal void SetAuthSuspect() => AuthSuspect = true;
 
         /// <summary>
         /// Creates a new <see cref="ConnectionMultiplexer"/> instance.
@@ -815,12 +575,9 @@ namespace StackExchange.Redis
         {
             SocketConnection.AssertDependencies();
 
-            if (configuration?.IsSentinel == true)
-            {
-                return SentinelPrimaryConnectAsync(configuration, log);
-            }
-
-            return ConnectImplAsync(PrepareConfig(configuration), log);
+            return configuration?.IsSentinel == true
+                ? SentinelPrimaryConnectAsync(configuration, log)
+                : ConnectImplAsync(PrepareConfig(configuration), log);
         }
 
         private static async Task<ConnectionMultiplexer> ConnectImplAsync(ConfigurationOptions configuration, TextWriter log = null)
@@ -828,95 +585,51 @@ namespace StackExchange.Redis
             IDisposable killMe = null;
             EventHandler<ConnectionFailedEventArgs> connectHandler = null;
             ConnectionMultiplexer muxer = null;
-            using (var logProxy = LogProxy.TryCreate(log))
+            using var logProxy = LogProxy.TryCreate(log);
+            try
             {
-                try
+                var sw = ValueStopwatch.StartNew();
+                logProxy?.WriteLine($"Connecting (async) on {RuntimeInformation.FrameworkDescription} (StackExchange.Redis: v{Utils.GetLibVersion()})");
+
+                muxer = CreateMultiplexer(configuration, logProxy, out connectHandler);
+                killMe = muxer;
+                Interlocked.Increment(ref muxer._connectAttemptCount);
+                bool configured = await muxer.ReconfigureAsync(first: true, reconfigureAll: false, logProxy, null, "connect").ObserveErrors().ForAwait();
+                if (!configured)
                 {
-                    var sw = ValueStopwatch.StartNew();
-                    logProxy?.WriteLine($"Connecting (async) on {RuntimeInformation.FrameworkDescription} (StackExchange.Redis: v{Utils.GetLibVersion()})");
-
-                    muxer = CreateMultiplexer(configuration, logProxy, out connectHandler);
-                    killMe = muxer;
-                    Interlocked.Increment(ref muxer._connectAttemptCount);
-                    bool configured = await muxer.ReconfigureAsync(first: true, reconfigureAll: false, logProxy, null, "connect").ObserveErrors().ForAwait();
-                    if (!configured)
-                    {
-                        throw ExceptionFactory.UnableToConnect(muxer, muxer.failureMessage);
-                    }
-                    killMe = null;
-                    Interlocked.Increment(ref muxer._connectCompletedCount);
-
-                    if (muxer.ServerSelectionStrategy.ServerType == ServerType.Sentinel)
-                    {
-                        // Initialize the Sentinel handlers
-                        muxer.InitializeSentinel(logProxy);
-                    }
-
-                    await configuration.AfterConnectAsync(muxer, logProxy != null ? logProxy.WriteLine : LogProxy.NullWriter).ForAwait();
-
-                    logProxy?.WriteLine($"Total connect time: {sw.ElapsedMilliseconds:n0} ms");
-
-                    return muxer;
+                    throw ExceptionFactory.UnableToConnect(muxer, muxer.failureMessage);
                 }
-                finally
+                killMe = null;
+                Interlocked.Increment(ref muxer._connectCompletedCount);
+
+                if (muxer.ServerSelectionStrategy.ServerType == ServerType.Sentinel)
                 {
-                    if (connectHandler != null) muxer.ConnectionFailed -= connectHandler;
-                    if (killMe != null) try { killMe.Dispose(); } catch { }
+                    // Initialize the Sentinel handlers
+                    muxer.InitializeSentinel(logProxy);
                 }
+
+                await configuration.AfterConnectAsync(muxer, logProxy != null ? logProxy.WriteLine : LogProxy.NullWriter).ForAwait();
+
+                logProxy?.WriteLine($"Total connect time: {sw.ElapsedMilliseconds:n0} ms");
+
+                return muxer;
+            }
+            finally
+            {
+                if (connectHandler != null) muxer.ConnectionFailed -= connectHandler;
+                if (killMe != null) try { killMe.Dispose(); } catch { }
             }
         }
 
         internal static ConfigurationOptions PrepareConfig(ConfigurationOptions config, bool sentinel = false)
         {
-            if (config == null)
-            {
-                throw new ArgumentNullException(nameof(config));
-            }
+            _ = config ?? throw new ArgumentNullException(nameof(config));
             if (config.EndPoints.Count == 0)
             {
                 throw new ArgumentException("No endpoints specified", nameof(config));
             }
 
-            // TODO: Switch to partial clone
-            config = config.Clone();
-
-            if (sentinel)
-            {
-                config.SetSentinelDefaults();
-                return config;
-            }
-            else
-            {
-                config.SetDefaultPorts();
-                return config;
-            }
-        }
-        private static ConnectionMultiplexer CreateMultiplexer(ConfigurationOptions configuration, LogProxy log, out EventHandler<ConnectionFailedEventArgs> connectHandler)
-        {
-            var muxer = new ConnectionMultiplexer(configuration);
-            connectHandler = null;
-            if (log != null)
-            {
-                // Create a detachable event-handler to log detailed errors if something happens during connect/handshake
-                connectHandler = (_, a) =>
-                {
-                    try
-                    {
-                        lock (log.SyncLock) // Keep the outer and any inner errors contiguous
-                        {
-                            var ex = a.Exception;
-                            log?.WriteLine($"Connection failed: {Format.ToString(a.EndPoint)} ({a.ConnectionType}, {a.FailureType}): {ex?.Message ?? "(unknown)"}");
-                            while ((ex = ex.InnerException) != null)
-                            {
-                                log?.WriteLine($"> {ex.Message}");
-                            }
-                        }
-                    }
-                    catch { }
-                };
-                muxer.ConnectionFailed += connectHandler;
-            }
-            return muxer;
+            return config.Clone().WithDefaults(sentinel);
         }
 
         /// <summary>
@@ -946,92 +659,9 @@ namespace StackExchange.Redis
         {
             SocketConnection.AssertDependencies();
 
-            if (configuration?.IsSentinel == true)
-            {
-                return SentinelPrimaryConnect(configuration, log);
-            }
-
-            return ConnectImpl(PrepareConfig(configuration), log);
-        }
-
-        /// <summary>
-        /// Create a new <see cref="ConnectionMultiplexer"/> instance that connects to a Sentinel server.
-        /// </summary>
-        /// <param name="configuration">The string configuration to use for this multiplexer.</param>
-        /// <param name="log">The <see cref="TextWriter"/> to log to.</param>
-        public static ConnectionMultiplexer SentinelConnect(string configuration, TextWriter log = null)
-        {
-            SocketConnection.AssertDependencies();
-            var config = ConfigurationOptions.Parse(configuration);
-            return ConnectImpl(PrepareConfig(config, sentinel: true), log);
-        }
-
-        /// <summary>
-        /// Create a new <see cref="ConnectionMultiplexer"/> instance that connects to a Sentinel server.
-        /// </summary>
-        /// <param name="configuration">The string configuration to use for this multiplexer.</param>
-        /// <param name="log">The <see cref="TextWriter"/> to log to.</param>
-        public static Task<ConnectionMultiplexer> SentinelConnectAsync(string configuration, TextWriter log = null)
-        {
-            SocketConnection.AssertDependencies();
-            var config = ConfigurationOptions.Parse(configuration);
-            return ConnectImplAsync(PrepareConfig(config, sentinel: true), log);
-        }
-
-        /// <summary>
-        /// Create a new <see cref="ConnectionMultiplexer"/> instance that connects to a Sentinel server.
-        /// </summary>
-        /// <param name="configuration">The configuration options to use for this multiplexer.</param>
-        /// <param name="log">The <see cref="TextWriter"/> to log to.</param>
-        public static ConnectionMultiplexer SentinelConnect(ConfigurationOptions configuration, TextWriter log = null)
-        {
-            SocketConnection.AssertDependencies();
-            return ConnectImpl(PrepareConfig(configuration, sentinel: true), log);
-        }
-
-        /// <summary>
-        /// Create a new <see cref="ConnectionMultiplexer"/> instance that connects to a Sentinel server.
-        /// </summary>
-        /// <param name="configuration">The configuration options to use for this multiplexer.</param>
-        /// <param name="log">The <see cref="TextWriter"/> to log to.</param>
-        public static Task<ConnectionMultiplexer> SentinelConnectAsync(ConfigurationOptions configuration, TextWriter log = null)
-        {
-            SocketConnection.AssertDependencies();
-            return ConnectImplAsync(PrepareConfig(configuration, sentinel: true), log);
-        }
-
-        /// <summary>
-        /// Create a new <see cref="ConnectionMultiplexer"/> instance that connects to a sentinel server, discovers the current primary server
-        /// for the specified <see cref="ConfigurationOptions.ServiceName"/> in the config and returns a managed connection to the current primary server.
-        /// </summary>
-        /// <param name="configuration">The configuration options to use for this multiplexer.</param>
-        /// <param name="log">The <see cref="TextWriter"/> to log to.</param>
-        private static ConnectionMultiplexer SentinelPrimaryConnect(ConfigurationOptions configuration, TextWriter log = null)
-        {
-            var sentinelConnection = SentinelConnect(configuration, log);
-
-            var muxer = sentinelConnection.GetSentinelMasterConnection(configuration, log);
-            // Set reference to sentinel connection so that we can dispose it
-            muxer.sentinelConnection = sentinelConnection;
-
-            return muxer;
-        }
-
-        /// <summary>
-        /// Create a new <see cref="ConnectionMultiplexer"/> instance that connects to a sentinel server, discovers the current primary server
-        /// for the specified <see cref="ConfigurationOptions.ServiceName"/> in the config and returns a managed connection to the current primary server.
-        /// </summary>
-        /// <param name="configuration">The configuration options to use for this multiplexer.</param>
-        /// <param name="log">The <see cref="TextWriter"/> to log to.</param>
-        private static async Task<ConnectionMultiplexer> SentinelPrimaryConnectAsync(ConfigurationOptions configuration, TextWriter log = null)
-        {
-            var sentinelConnection = await SentinelConnectAsync(configuration, log).ForAwait();
-
-            var muxer = sentinelConnection.GetSentinelMasterConnection(configuration, log);
-            // Set reference to sentinel connection so that we can dispose it
-            muxer.sentinelConnection = sentinelConnection;
-
-            return muxer;
+            return configuration?.IsSentinel == true
+                ? SentinelPrimaryConnect(configuration, log)
+                : ConnectImpl(PrepareConfig(configuration), log);
         }
 
         private static ConnectionMultiplexer ConnectImpl(ConfigurationOptions configuration, TextWriter log)
@@ -1039,59 +669,53 @@ namespace StackExchange.Redis
             IDisposable killMe = null;
             EventHandler<ConnectionFailedEventArgs> connectHandler = null;
             ConnectionMultiplexer muxer = null;
-            using (var logProxy = LogProxy.TryCreate(log))
+            using var logProxy = LogProxy.TryCreate(log);
+            try
             {
-                try
+                var sw = ValueStopwatch.StartNew();
+                logProxy?.WriteLine($"Connecting (sync) on {RuntimeInformation.FrameworkDescription} (StackExchange.Redis: v{Utils.GetLibVersion()})");
+
+                muxer = CreateMultiplexer(configuration, logProxy, out connectHandler);
+                killMe = muxer;
+                Interlocked.Increment(ref muxer._connectAttemptCount);
+                // note that task has timeouts internally, so it might take *just over* the regular timeout
+                var task = muxer.ReconfigureAsync(first: true, reconfigureAll: false, logProxy, null, "connect");
+
+                if (!task.Wait(muxer.SyncConnectTimeout(true)))
                 {
-                    var sw = ValueStopwatch.StartNew();
-                    logProxy?.WriteLine($"Connecting (sync) on {RuntimeInformation.FrameworkDescription} (StackExchange.Redis: v{Utils.GetLibVersion()})");
-
-                    muxer = CreateMultiplexer(configuration, logProxy, out connectHandler);
-                    killMe = muxer;
-                    Interlocked.Increment(ref muxer._connectAttemptCount);
-                    // note that task has timeouts internally, so it might take *just over* the regular timeout
-                    var task = muxer.ReconfigureAsync(first: true, reconfigureAll: false, logProxy, null, "connect");
-
-                    if (!task.Wait(muxer.SyncConnectTimeout(true)))
+                    task.ObserveErrors();
+                    if (muxer.RawConfig.AbortOnConnectFail)
                     {
-                        task.ObserveErrors();
-                        if (muxer.RawConfig.AbortOnConnectFail)
-                        {
-                            throw ExceptionFactory.UnableToConnect(muxer, "ConnectTimeout");
-                        }
-                        else
-                        {
-                            muxer.LastException = ExceptionFactory.UnableToConnect(muxer, "ConnectTimeout");
-                        }
+                        throw ExceptionFactory.UnableToConnect(muxer, "ConnectTimeout");
                     }
-
-                    if (!task.Result) throw ExceptionFactory.UnableToConnect(muxer, muxer.failureMessage);
-                    killMe = null;
-                    Interlocked.Increment(ref muxer._connectCompletedCount);
-
-                    if (muxer.ServerSelectionStrategy.ServerType == ServerType.Sentinel)
+                    else
                     {
-                        // Initialize the Sentinel handlers
-                        muxer.InitializeSentinel(logProxy);
+                        muxer.LastException = ExceptionFactory.UnableToConnect(muxer, "ConnectTimeout");
                     }
-
-                    configuration.AfterConnectAsync(muxer, logProxy != null ? logProxy.WriteLine : LogProxy.NullWriter).Wait(muxer.SyncConnectTimeout(true));
-
-                    logProxy?.WriteLine($"Total connect time: {sw.ElapsedMilliseconds:n0} ms");
-
-                    return muxer;
                 }
-                finally
+
+                if (!task.Result) throw ExceptionFactory.UnableToConnect(muxer, muxer.failureMessage);
+                killMe = null;
+                Interlocked.Increment(ref muxer._connectCompletedCount);
+
+                if (muxer.ServerSelectionStrategy.ServerType == ServerType.Sentinel)
                 {
-                    if (connectHandler != null && muxer != null) muxer.ConnectionFailed -= connectHandler;
-                    if (killMe != null) try { killMe.Dispose(); } catch { }
+                    // Initialize the Sentinel handlers
+                    muxer.InitializeSentinel(logProxy);
                 }
+
+                configuration.AfterConnectAsync(muxer, logProxy != null ? logProxy.WriteLine : LogProxy.NullWriter).Wait(muxer.SyncConnectTimeout(true));
+
+                logProxy?.WriteLine($"Total connect time: {sw.ElapsedMilliseconds:n0} ms");
+
+                return muxer;
+            }
+            finally
+            {
+                if (connectHandler != null && muxer != null) muxer.ConnectionFailed -= connectHandler;
+                if (killMe != null) try { killMe.Dispose(); } catch { }
             }
         }
-
-        private string failureMessage;
-        private readonly Hashtable servers = new Hashtable();
-        private volatile ServerSnapshot _serverSnapshot = ServerSnapshot.Empty;
 
         ReadOnlySpan<ServerEndPoint> IInternalConnectionMultiplexer.GetServerSnapshot() => GetServerSnapshot();
         internal ReadOnlySpan<ServerEndPoint> GetServerSnapshot() => _serverSnapshot.Span;
@@ -1109,7 +733,10 @@ namespace StackExchange.Redis
 
             internal ServerSnapshot Add(ServerEndPoint value)
             {
-                if (value == null) return this;
+                if (value == null)
+                {
+                    return this;
+                }
 
                 ServerEndPoint[] arr;
                 if (_arr.Length > _count)
@@ -1143,9 +770,11 @@ namespace StackExchange.Redis
 
         internal ServerEndPoint GetServerEndPoint(EndPoint endpoint, LogProxy log = null, bool activate = true)
         {
-            if (endpoint == null) return null;
-            var server = (ServerEndPoint)servers[endpoint];
-            if (server == null)
+            if (endpoint == null)
+            {
+                return null;
+            }
+            if (servers[endpoint] is not ServerEndPoint server)
             {
                 bool isNew = false;
                 lock (servers)
@@ -1153,8 +782,10 @@ namespace StackExchange.Redis
                     server = (ServerEndPoint)servers[endpoint];
                     if (server == null)
                     {
-                        if (_isDisposed) throw new ObjectDisposedException(ToString());
-
+                        if (_isDisposed)
+                        {
+                            throw new ObjectDisposedException(ToString());
+                        }
                         server = new ServerEndPoint(this, endpoint);
                         servers.Add(endpoint, server);
                         isNew = true;
@@ -1167,41 +798,6 @@ namespace StackExchange.Redis
             return server;
         }
 
-        internal readonly CommandMap CommandMap;
-
-        private ConnectionMultiplexer(ConfigurationOptions configuration)
-        {
-            IncludeDetailInExceptions = true;
-            IncludePerformanceCountersInExceptions = false;
-
-            RawConfig = configuration ?? throw new ArgumentNullException(nameof(configuration));
-
-            var map = CommandMap = configuration.CommandMap;
-            if (!string.IsNullOrWhiteSpace(configuration.Password)) map.AssertAvailable(RedisCommand.AUTH);
-
-            if (!map.IsAvailable(RedisCommand.ECHO) && !map.IsAvailable(RedisCommand.PING) && !map.IsAvailable(RedisCommand.TIME))
-            { // I mean really, give me a CHANCE! I need *something* to check the server is available to me...
-                // see also: SendTracer (matching logic)
-                map.AssertAvailable(RedisCommand.EXISTS);
-            }
-
-            TimeoutMilliseconds = configuration.SyncTimeout;
-            AsyncTimeoutMilliseconds = configuration.AsyncTimeout;
-
-            OnCreateReaderWriter(configuration);
-            ServerSelectionStrategy = new ServerSelectionStrategy(this);
-
-            var configChannel = configuration.ConfigurationChannel;
-            if (!string.IsNullOrWhiteSpace(configChannel))
-            {
-                ConfigurationChangedChannel = Encoding.UTF8.GetBytes(configChannel);
-            }
-            lastHeartbeatTicks = Environment.TickCount;
-        }
-
-        partial void OnCreateReaderWriter(ConfigurationOptions configuration);
-
-        internal const int MillisecondsPerHeartbeat = 1000;
         private sealed class TimerToken
         {
             public TimerToken(ConnectionMultiplexer muxer)
@@ -1238,7 +834,6 @@ namespace StackExchange.Redis
             }
         }
 
-        private int _activeHeartbeatErrors;
         private void OnHeartbeat()
         {
             try
@@ -1267,21 +862,6 @@ namespace StackExchange.Redis
                 }
             }
         }
-
-        private int lastHeartbeatTicks;
-        private static int lastGlobalHeartbeatTicks = Environment.TickCount;
-        internal long LastHeartbeatSecondsAgo
-        {
-            get
-            {
-                if (pulse == null) return -1;
-                return unchecked(Environment.TickCount - Thread.VolatileRead(ref lastHeartbeatTicks)) / 1000;
-            }
-        }
-
-        internal Exception LastException { get; set; }
-
-        internal static long LastGlobalHeartbeatSecondsAgo => unchecked(Environment.TickCount - Thread.VolatileRead(ref lastGlobalHeartbeatTicks)) / 1000;
 
         /// <summary>
         /// Obtain a pub/sub subscriber connection to the specified server.
@@ -1351,19 +931,66 @@ namespace StackExchange.Redis
         }
 
         /// <summary>
+        /// Compute the hash-slot of a specified key.
+        /// </summary>
+        /// <param name="key">The key to get a hash slot ID for.</param>
+        public int HashSlot(RedisKey key) => ServerSelectionStrategy.HashSlot(key);
+
+        internal ServerEndPoint AnyServer(ServerType serverType, uint startOffset, RedisCommand command, CommandFlags flags, bool allowDisconnected)
+        {
+            var tmp = GetServerSnapshot();
+            int len = tmp.Length;
+            ServerEndPoint fallback = null;
+            for (int i = 0; i < len; i++)
+            {
+                var server = tmp[(int)(((uint)i + startOffset) % len)];
+                if (server != null && server.ServerType == serverType && server.IsSelectable(command, allowDisconnected))
+                {
+                    if (server.IsReplica)
+                    {
+                        switch (flags)
+                        {
+                            case CommandFlags.DemandReplica:
+                            case CommandFlags.PreferReplica:
+                                return server;
+                            case CommandFlags.PreferMaster:
+                                fallback = server;
+                                break;
+                        }
+                    }
+                    else
+                    {
+                        switch (flags)
+                        {
+                            case CommandFlags.DemandMaster:
+                            case CommandFlags.PreferMaster:
+                                return server;
+                            case CommandFlags.PreferReplica:
+                                fallback = server;
+                                break;
+                        }
+                    }
+                }
+            }
+            return fallback;
+        }
+
+        /// <summary>
         /// Obtain a configuration API for an individual server.
         /// </summary>
         /// <param name="host">The host to get a server for.</param>
         /// <param name="port">The port for <paramref name="host"/> to get a server for.</param>
         /// <param name="asyncState">The async state to pass into the resulting <see cref="RedisServer"/>.</param>
-        public IServer GetServer(string host, int port, object asyncState = null) => GetServer(Format.ParseEndPoint(host, port), asyncState);
+        public IServer GetServer(string host, int port, object asyncState = null) =>
+            GetServer(Format.ParseEndPoint(host, port), asyncState);
 
         /// <summary>
         /// Obtain a configuration API for an individual server.
         /// </summary>
         /// <param name="hostAndPort">The "host:port" string to get a server for.</param>
         /// <param name="asyncState">The async state to pass into the resulting <see cref="RedisServer"/>.</param>
-        public IServer GetServer(string hostAndPort, object asyncState = null) => GetServer(Format.TryParseEndPoint(hostAndPort), asyncState);
+        public IServer GetServer(string hostAndPort, object asyncState = null) =>
+            GetServer(Format.TryParseEndPoint(hostAndPort), asyncState);
 
         /// <summary>
         /// Obtain a configuration API for an individual server.
@@ -1379,15 +1006,21 @@ namespace StackExchange.Redis
         /// <param name="asyncState">The async state to pass into the resulting <see cref="RedisServer"/>.</param>
         public IServer GetServer(EndPoint endpoint, object asyncState = null)
         {
-            if (endpoint == null) throw new ArgumentNullException(nameof(endpoint));
+            _ = endpoint ?? throw new ArgumentNullException(nameof(endpoint));
             if (!RawConfig.Proxy.SupportsServerApi())
             {
                 throw new NotSupportedException($"The server API is not available via {RawConfig.Proxy}");
             }
-            var server = (ServerEndPoint)servers[endpoint];
-            if (server == null) throw new ArgumentException("The specified endpoint is not defined", nameof(endpoint));
+            var server = servers[endpoint] as ServerEndPoint ?? throw new ArgumentException("The specified endpoint is not defined", nameof(endpoint));
             return new RedisServer(this, server, asyncState);
         }
+
+        /// <summary>
+        /// Get the hash-slot associated with a given key, if applicable.
+        /// This can be useful for grouping operations.
+        /// </summary>
+        /// <param name="key">The <see cref="RedisKey"/> to determine the hash slot for.</param>
+        public int GetHashSlot(RedisKey key) => ServerSelectionStrategy.HashSlot(key);
 
         /// <summary>
         /// The number of operations that have been performed on all connections.
@@ -1403,41 +1036,6 @@ namespace StackExchange.Redis
             }
         }
 
-        private string activeConfigCause;
-
-        internal bool ReconfigureIfNeeded(EndPoint blame, bool fromBroadcast, string cause, bool publishReconfigure = false, CommandFlags flags = CommandFlags.None)
-        {
-            if (fromBroadcast)
-            {
-                OnConfigurationChangedBroadcast(blame);
-            }
-            string activeCause = Volatile.Read(ref activeConfigCause);
-            if (activeCause == null)
-            {
-                bool reconfigureAll = fromBroadcast || publishReconfigure;
-                Trace("Configuration change detected; checking nodes", "Configuration");
-                ReconfigureAsync(first: false, reconfigureAll, null, blame, cause, publishReconfigure, flags).ObserveErrors();
-                return true;
-            }
-            else
-            {
-                Trace("Configuration change skipped; already in progress via " + activeCause, "Configuration");
-                return false;
-            }
-        }
-
-        /// <summary>
-        /// Reconfigure the current connections based on the existing configuration.
-        /// </summary>
-        /// <param name="log">The <see cref="TextWriter"/> to log to.</param>
-        public async Task<bool> ConfigureAsync(TextWriter log = null)
-        {
-            using (var logProxy = LogProxy.TryCreate(log))
-            {
-                return await ReconfigureAsync(first: false, reconfigureAll: true, logProxy, null, "configure").ObserveErrors();
-            }
-        }
-
         /// <summary>
         /// Reconfigure the current connections based on the existing configuration.
         /// </summary>
@@ -1446,24 +1044,32 @@ namespace StackExchange.Redis
         {
             // Note we expect ReconfigureAsync to internally allow [n] duration,
             // so to avoid near misses, here we wait 2*[n].
-            using (var logProxy = LogProxy.TryCreate(log))
+            using var logProxy = LogProxy.TryCreate(log);
+            var task = ReconfigureAsync(first: false, reconfigureAll: true, logProxy, null, "configure");
+            if (!task.Wait(SyncConnectTimeout(false)))
             {
-                var task = ReconfigureAsync(first: false, reconfigureAll: true, logProxy, null, "configure");
-                if (!task.Wait(SyncConnectTimeout(false)))
+                task.ObserveErrors();
+                if (RawConfig.AbortOnConnectFail)
                 {
-                    task.ObserveErrors();
-                    if (RawConfig.AbortOnConnectFail)
-                    {
-                        throw new TimeoutException();
-                    }
-                    else
-                    {
-                        LastException = new TimeoutException("ConnectTimeout");
-                    }
-                    return false;
+                    throw new TimeoutException();
                 }
-                return task.Result;
+                else
+                {
+                    LastException = new TimeoutException("ConnectTimeout");
+                }
+                return false;
             }
+            return task.Result;
+        }
+
+        /// <summary>
+        /// Reconfigure the current connections based on the existing configuration.
+        /// </summary>
+        /// <param name="log">The <see cref="TextWriter"/> to log to.</param>
+        public async Task<bool> ConfigureAsync(TextWriter log = null)
+        {
+            using var logProxy = LogProxy.TryCreate(log);
+            return await ReconfigureAsync(first: false, reconfigureAll: true, logProxy, null, "configure").ObserveErrors();
         }
 
         internal int SyncConnectTimeout(bool forConnect)
@@ -1484,11 +1090,9 @@ namespace StackExchange.Redis
         /// </summary>
         public string GetStatus()
         {
-            using (var sw = new StringWriter())
-            {
-                GetStatus(sw);
-                return sw.ToString();
-            }
+            using var sw = new StringWriter();
+            GetStatus(sw);
+            return sw.ToString();
         }
 
         /// <summary>
@@ -1497,10 +1101,8 @@ namespace StackExchange.Redis
         /// <param name="log">The <see cref="TextWriter"/> to log to.</param>
         public void GetStatus(TextWriter log)
         {
-            using (var proxy = LogProxy.TryCreate(log))
-            {
-                GetStatus(proxy);
-            }
+            using var proxy = LogProxy.TryCreate(log);
+            GetStatus(proxy);
         }
 
         internal void GetStatus(LogProxy log)
@@ -1531,10 +1133,30 @@ namespace StackExchange.Redis
             }
         }
 
+        internal bool ReconfigureIfNeeded(EndPoint blame, bool fromBroadcast, string cause, bool publishReconfigure = false, CommandFlags flags = CommandFlags.None)
+        {
+            if (fromBroadcast)
+            {
+                OnConfigurationChangedBroadcast(blame);
+            }
+            string activeCause = Volatile.Read(ref activeConfigCause);
+            if (activeCause == null)
+            {
+                bool reconfigureAll = fromBroadcast || publishReconfigure;
+                Trace("Configuration change detected; checking nodes", "Configuration");
+                ReconfigureAsync(first: false, reconfigureAll, null, blame, cause, publishReconfigure, flags).ObserveErrors();
+                return true;
+            }
+            else
+            {
+                Trace("Configuration change skipped; already in progress via " + activeCause, "Configuration");
+                return false;
+            }
+        }
+
         /// <summary>
         /// Triggers a reconfigure of this multiplexer.
         /// This re-assessment of all server endpoints to get the current topology and adjust, the same as if we had first connected.
-        /// TODO: Naming?
         /// </summary>
         public Task<bool> ReconfigureAsync(string reason) =>
             ReconfigureAsync(first: false, reconfigureAll: false, log: null, blame: null, cause: reason);
@@ -1865,12 +1487,20 @@ namespace StackExchange.Redis
             finally
             {
                 Trace("Exiting reconfiguration...");
-                OnTraceLog(log);
                 if (ranThisCall) Interlocked.Exchange(ref activeConfigCause, null);
                 if (!first) OnConfigurationChanged(blame);
                 Trace("Reconfiguration exited");
             }
         }
+
+        /// <summary>
+        /// Gets all endpoints defined on the multiplexer.
+        /// </summary>
+        /// <param name="configuredOnly">Whether to get only the endpoints specified explicitly in the config.</param>
+        public EndPoint[] GetEndPoints(bool configuredOnly = false) =>
+            configuredOnly
+                ? RawConfig.EndPoints.ToArray()
+                : _serverSnapshot.GetEndPoints();
 
         private async Task<EndPointCollection> GetEndpointsFromClusterNodes(ServerEndPoint server, LogProxy log)
         {
@@ -1906,10 +1536,6 @@ namespace StackExchange.Redis
                 server.ResetNonConnected();
             }
         }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE0060:Remove unused parameter", Justification = "Used - it's a partial")]
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Performance", "CA1822:Mark members as static", Justification = "Partial - may use instance data")]
-        partial void OnTraceLog(LogProxy log, [CallerMemberName] string caller = null);
 
         private static ServerEndPoint NominatePreferredPrimary(LogProxy log, ServerEndPoint[] servers, bool useTieBreakers, List<ServerEndPoint> primaries)
         {
@@ -2052,19 +1678,22 @@ namespace StackExchange.Redis
 
         internal void UpdateClusterRange(ClusterConfiguration configuration)
         {
-            if (configuration == null) return;
+            if (configuration is null)
+            {
+                return;
+            }
             foreach (var node in configuration.Nodes)
             {
                 if (node.IsReplica || node.Slots.Count == 0) continue;
                 foreach (var slot in node.Slots)
                 {
-                    var server = GetServerEndPoint(node.EndPoint);
-                    if (server != null) ServerSelectionStrategy.UpdateClusterRange(slot.From, slot.To, server);
+                    if (GetServerEndPoint(node.EndPoint) is ServerEndPoint server)
+                    {
+                        ServerSelectionStrategy.UpdateClusterRange(slot.From, slot.To, server);
+                    }
                 }
             }
         }
-
-        private IDisposable pulse;
 
         internal ServerEndPoint SelectServer(Message message) =>
             message == null ? null : ServerSelectionStrategy.Select(message);
@@ -2138,6 +1767,7 @@ namespace StackExchange.Redis
             Trace("No server or server unavailable - aborting: " + message);
             return false;
         }
+
         private ValueTask<WriteResult> TryPushMessageToBridgeAsync<T>(Message message, ResultProcessor<T> processor, IResultBox<T> resultBox, ref ServerEndPoint server)
             => PrepareToPushMessageToBridge(message, processor, resultBox, ref server) ? server.TryWriteAsync(message) : new ValueTask<WriteResult>(WriteResult.NoConnectionAvailable);
 
@@ -2148,526 +1778,7 @@ namespace StackExchange.Redis
         /// <summary>
         /// Gets the client name for this multiplexer.
         /// </summary>
-        public override string ToString()
-        {
-            string s = ClientName;
-            if (string.IsNullOrWhiteSpace(s)) s = GetType().Name;
-            return s;
-        }
-
-        internal readonly byte[] ConfigurationChangedChannel; // This gets accessed for every received event; let's make sure we can process it "raw"
-        internal readonly byte[] UniqueId = Guid.NewGuid().ToByteArray(); // Unique identifier used when tracing
-
-        /// <summary>
-        /// Gets or sets whether asynchronous operations should be invoked in a way that guarantees their original delivery order.
-        /// </summary>
-        [Obsolete("Not supported; if you require ordered pub/sub, please see " + nameof(ChannelMessageQueue) + ", will be removed in 3.0", false)]
-        public bool PreserveAsyncOrder
-        {
-            get => false;
-            set { }
-        }
-
-        /// <summary>
-        /// Indicates whether any servers are connected.
-        /// </summary>
-        public bool IsConnected
-        {
-            get
-            {
-                var tmp = GetServerSnapshot();
-                for (int i = 0; i < tmp.Length; i++)
-                    if (tmp[i].IsConnected) return true;
-                return false;
-            }
-        }
-
-        /// <summary>
-        /// Indicates whether any servers are currently trying to connect.
-        /// </summary>
-        public bool IsConnecting
-        {
-            get
-            {
-                var tmp = GetServerSnapshot();
-                for (int i = 0; i < tmp.Length; i++)
-                    if (tmp[i].IsConnecting) return true;
-                return false;
-            }
-        }
-
-        internal ConfigurationOptions RawConfig { get; }
-
-        internal ServerSelectionStrategy ServerSelectionStrategy { get; }
-
-        internal Timer sentinelPrimaryReconnectTimer;
-
-        internal Dictionary<string, ConnectionMultiplexer> sentinelConnectionChildren = new Dictionary<string, ConnectionMultiplexer>();
-        internal ConnectionMultiplexer sentinelConnection = null;
-
-        /// <summary>
-        /// Initializes the connection as a Sentinel connection and adds the necessary event handlers to track changes to the managed primaries.
-        /// </summary>
-        /// <param name="logProxy">The writer to log to, if any.</param>
-        internal void InitializeSentinel(LogProxy logProxy)
-        {
-            if (ServerSelectionStrategy.ServerType != ServerType.Sentinel)
-            {
-                return;
-            }
-
-            // Subscribe to sentinel change events
-            ISubscriber sub = GetSubscriber();
-
-            if (sub.SubscribedEndpoint("+switch-master") == null)
-            {
-                sub.Subscribe("+switch-master", (_, message) =>
-                {
-                    string[] messageParts = ((string)message).Split(new[] { ' ' }, StringSplitOptions.RemoveEmptyEntries);
-                    EndPoint switchBlame = Format.TryParseEndPoint(string.Format("{0}:{1}", messageParts[1], messageParts[2]));
-
-                    lock (sentinelConnectionChildren)
-                    {
-                        // Switch the primary if we have connections for that service
-                        if (sentinelConnectionChildren.ContainsKey(messageParts[0]))
-                        {
-                            ConnectionMultiplexer child = sentinelConnectionChildren[messageParts[0]];
-
-                            // Is the connection still valid?
-                            if (child.IsDisposed)
-                            {
-                                child.ConnectionFailed -= OnManagedConnectionFailed;
-                                child.ConnectionRestored -= OnManagedConnectionRestored;
-                                sentinelConnectionChildren.Remove(messageParts[0]);
-                            }
-                            else
-                            {
-                                SwitchPrimary(switchBlame, sentinelConnectionChildren[messageParts[0]]);
-                            }
-                        }
-                    }
-                }, CommandFlags.FireAndForget);
-            }
-
-            // If we lose connection to a sentinel server,
-            // we need to reconfigure to make sure we still have a subscription to the +switch-master channel
-            ConnectionFailed += (sender, e) =>
-            {
-                // Reconfigure to get subscriptions back online
-                ReconfigureAsync(first: false, reconfigureAll: true, logProxy, e.EndPoint, "Lost sentinel connection", false).Wait();
-            };
-
-            // Subscribe to new sentinels being added
-            if (sub.SubscribedEndpoint("+sentinel") == null)
-            {
-                sub.Subscribe("+sentinel", (_, message) =>
-                {
-                    string[] messageParts = ((string)message).Split(new[] { ' ' }, StringSplitOptions.RemoveEmptyEntries);
-                    UpdateSentinelAddressList(messageParts[0]);
-                }, CommandFlags.FireAndForget);
-            }
-        }
-
-        /// <summary>
-        /// Returns a managed connection to the primary server indicated by the <see cref="ConfigurationOptions.ServiceName"/> in the config.
-        /// </summary>
-        /// <param name="config">The configuration to be used when connecting to the primary.</param>
-        /// <param name="log">The writer to log to, if any.</param>
-        public ConnectionMultiplexer GetSentinelMasterConnection(ConfigurationOptions config, TextWriter log = null)
-        {
-            if (ServerSelectionStrategy.ServerType != ServerType.Sentinel)
-            {
-                throw new RedisConnectionException(ConnectionFailureType.UnableToConnect,
-                    "Sentinel: The ConnectionMultiplexer is not a Sentinel connection. Detected as: " + ServerSelectionStrategy.ServerType);
-            }
-
-            if (string.IsNullOrEmpty(config.ServiceName))
-                throw new ArgumentException("A ServiceName must be specified.");
-
-            lock (sentinelConnectionChildren)
-            {
-                if (sentinelConnectionChildren.TryGetValue(config.ServiceName, out var sentinelConnectionChild) && !sentinelConnectionChild.IsDisposed)
-                    return sentinelConnectionChild;
-            }
-
-            bool success = false;
-            ConnectionMultiplexer connection = null;
-
-            var sw = ValueStopwatch.StartNew();
-            do
-            {
-                // Get an initial endpoint - try twice
-                EndPoint newPrimaryEndPoint = GetConfiguredPrimaryForService(config.ServiceName)
-                                             ?? GetConfiguredPrimaryForService(config.ServiceName);
-
-                if (newPrimaryEndPoint == null)
-                {
-                    throw new RedisConnectionException(ConnectionFailureType.UnableToConnect,
-                        $"Sentinel: Failed connecting to configured primary for service: {config.ServiceName}");
-                }
-
-                EndPoint[] replicaEndPoints = GetReplicasForService(config.ServiceName)
-                                           ?? GetReplicasForService(config.ServiceName);
-
-                // Replace the primary endpoint, if we found another one
-                // If not, assume the last state is the best we have and minimize the race
-                if (config.EndPoints.Count == 1)
-                {
-                    config.EndPoints[0] = newPrimaryEndPoint;
-                }
-                else
-                {
-                    config.EndPoints.Clear();
-                    config.EndPoints.TryAdd(newPrimaryEndPoint);
-                }
-
-                foreach (var replicaEndPoint in replicaEndPoints)
-                {
-                    config.EndPoints.TryAdd(replicaEndPoint);
-                }
-
-                connection = ConnectImpl(config, log);
-
-                // verify role is primary according to:
-                // https://redis.io/topics/sentinel-clients
-                if (connection.GetServer(newPrimaryEndPoint)?.Role().Value == RedisLiterals.master)
-                {
-                    success = true;
-                    break;
-                }
-
-                Thread.Sleep(100);
-            } while (sw.ElapsedMilliseconds < config.ConnectTimeout);
-
-            if (!success)
-            {
-                throw new RedisConnectionException(ConnectionFailureType.UnableToConnect,
-                    $"Sentinel: Failed connecting to configured primary for service: {config.ServiceName}");
-            }
-
-            // Attach to reconnect event to ensure proper connection to the new primary
-            connection.ConnectionRestored += OnManagedConnectionRestored;
-
-            // If we lost the connection, run a switch to a least try and get updated info about the primary
-            connection.ConnectionFailed += OnManagedConnectionFailed;
-
-            lock (sentinelConnectionChildren)
-            {
-                sentinelConnectionChildren[connection.RawConfig.ServiceName] = connection;
-            }
-
-            // Perform the initial switchover
-            SwitchPrimary(RawConfig.EndPoints[0], connection, log);
-
-            return connection;
-        }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Roslynator", "RCS1075:Avoid empty catch clause that catches System.Exception.", Justification = "We don't care.")]
-        internal void OnManagedConnectionRestored(object sender, ConnectionFailedEventArgs e)
-        {
-            ConnectionMultiplexer connection = (ConnectionMultiplexer)sender;
-
-            var oldTimer = Interlocked.Exchange(ref connection.sentinelPrimaryReconnectTimer, null);
-            oldTimer?.Dispose();
-
-            try
-            {
-                // Run a switch to make sure we have update-to-date
-                // information about which primary we should connect to
-                SwitchPrimary(e.EndPoint, connection);
-
-                try
-                {
-                    // Verify that the reconnected endpoint is a primary,
-                    // and the correct one otherwise we should reconnect
-                    if (connection.GetServer(e.EndPoint).IsReplica || e.EndPoint != connection.currentSentinelPrimaryEndPoint)
-                    {
-                        // This isn't a primary, so try connecting again
-                        SwitchPrimary(e.EndPoint, connection);
-                    }
-                }
-                catch (Exception)
-                {
-                    // If we get here it means that we tried to reconnect to a server that is no longer
-                    // considered a primary by Sentinel and was removed from the list of endpoints.
-
-                    // If we caught an exception, we may have gotten a stale endpoint
-                    // we are not aware of, so retry
-                    SwitchPrimary(e.EndPoint, connection);
-                }
-            }
-            catch (Exception)
-            {
-                // Log, but don't throw in an event handler
-                // TODO: Log via new event handler? a la ConnectionFailed?
-            }
-        }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Roslynator", "RCS1075:Avoid empty catch clause that catches System.Exception.", Justification = "We don't care.")]
-        internal void OnManagedConnectionFailed(object sender, ConnectionFailedEventArgs e)
-        {
-            ConnectionMultiplexer connection = (ConnectionMultiplexer)sender;
-            // Periodically check to see if we can reconnect to the proper primary.
-            // This is here in case we lost our subscription to a good sentinel instance
-            // or if we miss the published primary change.
-            if (connection.sentinelPrimaryReconnectTimer == null)
-            {
-                connection.sentinelPrimaryReconnectTimer = new Timer(_ =>
-                {
-                    try
-                    {
-                        // Attempt, but do not fail here
-                        SwitchPrimary(e.EndPoint, connection);
-                    }
-                    catch (Exception)
-                    {
-                    }
-                    finally
-                    {
-                        connection.sentinelPrimaryReconnectTimer?.Change(TimeSpan.FromSeconds(1), Timeout.InfiniteTimeSpan);
-                    }
-                }, null, TimeSpan.Zero, Timeout.InfiniteTimeSpan);
-            }
-        }
-
-        internal EndPoint GetConfiguredPrimaryForService(string serviceName) =>
-            GetServerSnapshot()
-                .ToArray()
-                .Where(s => s.ServerType == ServerType.Sentinel)
-                .AsParallel()
-                .Select(s =>
-                {
-                    try { return GetServer(s.EndPoint).SentinelGetMasterAddressByName(serviceName); }
-                    catch { return null; }
-                })
-                .FirstOrDefault(r => r != null);
-
-        internal EndPoint currentSentinelPrimaryEndPoint;
-
-        internal EndPoint[] GetReplicasForService(string serviceName) =>
-            GetServerSnapshot()
-                .ToArray()
-                .Where(s => s.ServerType == ServerType.Sentinel)
-                .AsParallel()
-                .Select(s =>
-                {
-                    try { return GetServer(s.EndPoint).SentinelGetReplicaAddresses(serviceName); }
-                    catch { return null; }
-                })
-                .FirstOrDefault(r => r != null);
-
-        /// <summary>
-        /// Switches the SentinelMasterConnection over to a new primary.
-        /// </summary>
-        /// <param name="switchBlame">The endpoint responsible for the switch.</param>
-        /// <param name="connection">The connection that should be switched over to a new primary endpoint.</param>
-        /// <param name="log">The writer to log to, if any.</param>
-        internal void SwitchPrimary(EndPoint switchBlame, ConnectionMultiplexer connection, TextWriter log = null)
-        {
-            if (log == null) log = TextWriter.Null;
-
-            using (var logProxy = LogProxy.TryCreate(log))
-            {
-                string serviceName = connection.RawConfig.ServiceName;
-
-                // Get new primary - try twice
-                EndPoint newPrimaryEndPoint = GetConfiguredPrimaryForService(serviceName)
-                                           ?? GetConfiguredPrimaryForService(serviceName)
-                                           ?? throw new RedisConnectionException(ConnectionFailureType.UnableToConnect,
-                                                $"Sentinel: Failed connecting to switch primary for service: {serviceName}");
-
-                connection.currentSentinelPrimaryEndPoint = newPrimaryEndPoint;
-
-                if (!connection.servers.Contains(newPrimaryEndPoint))
-                {
-                    EndPoint[] replicaEndPoints = GetReplicasForService(serviceName)
-                                               ?? GetReplicasForService(serviceName);
-
-                    connection.servers.Clear();
-                    connection.RawConfig.EndPoints.Clear();
-                    connection.RawConfig.EndPoints.TryAdd(newPrimaryEndPoint);
-                    foreach (var replicaEndPoint in replicaEndPoints)
-                    {
-                        connection.RawConfig.EndPoints.TryAdd(replicaEndPoint);
-                    }
-                    Trace($"Switching primary to {newPrimaryEndPoint}");
-                    // Trigger a reconfigure
-                    connection.ReconfigureAsync(first: false, reconfigureAll: false, logProxy, switchBlame,
-                        $"Primary switch {serviceName}", false, CommandFlags.PreferMaster).Wait();
-
-                    UpdateSentinelAddressList(serviceName);
-                }
-            }
-        }
-
-        internal void UpdateSentinelAddressList(string serviceName)
-        {
-            var firstCompleteRequest = GetServerSnapshot()
-                                        .ToArray()
-                                        .Where(s => s.ServerType == ServerType.Sentinel)
-                                        .AsParallel()
-                                        .Select(s =>
-                                        {
-                                            try { return GetServer(s.EndPoint).SentinelGetSentinelAddresses(serviceName); }
-                                            catch { return null; }
-                                        })
-                                        .FirstOrDefault(r => r != null);
-
-            // Ignore errors, as having an updated sentinel list is not essential
-            if (firstCompleteRequest == null)
-                return;
-
-            bool hasNew = false;
-            foreach (EndPoint newSentinel in firstCompleteRequest.Where(x => !RawConfig.EndPoints.Contains(x)))
-            {
-                hasNew = true;
-                RawConfig.EndPoints.TryAdd(newSentinel);
-            }
-
-            if (hasNew)
-            {
-                // Reconfigure the sentinel multiplexer if we added new endpoints
-                ReconfigureAsync(first: false, reconfigureAll: true, null, RawConfig.EndPoints[0], "Updating Sentinel List", false).Wait();
-            }
-        }
-
-        /// <summary>
-        /// Close all connections and release all resources associated with this object.
-        /// </summary>
-        /// <param name="allowCommandsToComplete">Whether to allow all in-queue commands to complete first.</param>
-        public void Close(bool allowCommandsToComplete = true)
-        {
-            if (_isDisposed) return;
-
-            OnClosing(false);
-            _isDisposed = true;
-            _profilingSessionProvider = null;
-            using (var tmp = pulse)
-            {
-                pulse = null;
-            }
-
-            if (allowCommandsToComplete)
-            {
-                var quits = QuitAllServers();
-                WaitAllIgnoreErrors(quits);
-            }
-            DisposeAndClearServers();
-            OnCloseReaderWriter();
-            OnClosing(true);
-            Interlocked.Increment(ref _connectionCloseCount);
-        }
-
-        partial void OnCloseReaderWriter();
-
-        private void DisposeAndClearServers()
-        {
-            lock (servers)
-            {
-                var iter = servers.GetEnumerator();
-                while (iter.MoveNext())
-                {
-                    var server = (ServerEndPoint)iter.Value;
-                    server.Dispose();
-                }
-                servers.Clear();
-            }
-        }
-
-        private Task[] QuitAllServers()
-        {
-            var quits = new Task[2 * servers.Count];
-            lock (servers)
-            {
-                var iter = servers.GetEnumerator();
-                int index = 0;
-                while (iter.MoveNext())
-                {
-                    var server = (ServerEndPoint)iter.Value;
-                    quits[index++] = server.Close(ConnectionType.Interactive);
-                    quits[index++] = server.Close(ConnectionType.Subscription);
-                }
-            }
-            return quits;
-        }
-
-        /// <summary>
-        /// Close all connections and release all resources associated with this object.
-        /// </summary>
-        /// <param name="allowCommandsToComplete">Whether to allow all in-queue commands to complete first.</param>
-        public async Task CloseAsync(bool allowCommandsToComplete = true)
-        {
-            _isDisposed = true;
-            using (var tmp = pulse)
-            {
-                pulse = null;
-            }
-
-            if (allowCommandsToComplete)
-            {
-                var quits = QuitAllServers();
-                await WaitAllIgnoreErrorsAsync("quit", quits, RawConfig.AsyncTimeout, null).ForAwait();
-            }
-
-            DisposeAndClearServers();
-        }
-
-        /// <summary>
-        /// Release all resources associated with this object.
-        /// </summary>
-        public void Dispose()
-        {
-            GC.SuppressFinalize(this);
-            Close(!_isDisposed);
-            sentinelConnection?.Dispose();
-            var oldTimer = Interlocked.Exchange(ref sentinelPrimaryReconnectTimer, null);
-            oldTimer?.Dispose();
-        }
-
-        internal Task<T> ExecuteAsyncImpl<T>(Message message, ResultProcessor<T> processor, object state, ServerEndPoint server)
-        {
-            if (_isDisposed) throw new ObjectDisposedException(ToString());
-
-            if (message == null)
-            {
-                return CompletedTask<T>.Default(state);
-            }
-
-            TaskCompletionSource<T> tcs = null;
-            IResultBox<T> source = null;
-            if (!message.IsFireAndForget)
-            {
-                source = TaskResultBox<T>.Create(out tcs, state);
-            }
-            var write = TryPushMessageToBridgeAsync(message, processor, source, ref server);
-            if (!write.IsCompletedSuccessfully) return ExecuteAsyncImpl_Awaited<T>(this, write, tcs, message, server);
-
-            if (tcs == null)
-            {
-                return CompletedTask<T>.Default(null); // F+F explicitly does not get async-state
-            }
-            else
-            {
-                var result = write.Result;
-                if (result != WriteResult.Success)
-                {
-                    var ex = GetException(result, message, server);
-                    ThrowFailed(tcs, ex);
-                }
-                return tcs.Task;
-            }
-        }
-
-        private static async Task<T> ExecuteAsyncImpl_Awaited<T>(ConnectionMultiplexer @this, ValueTask<WriteResult> write, TaskCompletionSource<T> tcs, Message message, ServerEndPoint server)
-        {
-            var result = await write.ForAwait();
-            if (result != WriteResult.Success)
-            {
-                var ex = @this.GetException(result, message, server);
-                ThrowFailed(tcs, ex);
-            }
-            return tcs == null ? default(T) : await tcs.Task.ForAwait();
-        }
+        public override string ToString() => string.IsNullOrWhiteSpace(ClientName) ? GetType().Name : ClientName;
 
         internal Exception GetException(WriteResult result, Message message, ServerEndPoint server) => result switch
         {
@@ -2743,39 +1854,50 @@ namespace StackExchange.Redis
             }
         }
 
-        /// <summary>
-        /// Should exceptions include identifiable details? (key names, additional .Data annotations)
-        /// </summary>
-        public bool IncludeDetailInExceptions { get; set; }
-
-        /// <summary>
-        /// Should exceptions include performance counter details? (CPU usage, etc - note that this can be problematic on some platforms)
-        /// </summary>
-        public bool IncludePerformanceCountersInExceptions { get; set; }
-
-        internal int haveStormLog = 0;
-        internal string stormLogSnapshot;
-        /// <summary>
-        /// Limit at which to start recording unusual busy patterns (only one log will be retained at a time).
-        /// Set to a negative value to disable this feature.
-        /// </summary>
-        public int StormLogThreshold { get; set; } = 15;
-
-        /// <summary>
-        /// Obtains the log of unusual busy patterns.
-        /// </summary>
-        public string GetStormLog() => Volatile.Read(ref stormLogSnapshot);
-
-        /// <summary>
-        /// Resets the log of unusual busy patterns.
-        /// </summary>
-        public void ResetStormLog()
+        internal Task<T> ExecuteAsyncImpl<T>(Message message, ResultProcessor<T> processor, object state, ServerEndPoint server)
         {
-            Interlocked.Exchange(ref stormLogSnapshot, null);
-            Interlocked.Exchange(ref haveStormLog, 0);
-        }
+            if (_isDisposed) throw new ObjectDisposedException(ToString());
 
-        private long syncTimeouts, fireAndForgets, asyncTimeouts;
+            if (message == null)
+            {
+                return CompletedTask<T>.Default(state);
+            }
+
+            static async Task<T> ExecuteAsyncImpl_Awaited(ConnectionMultiplexer @this, ValueTask<WriteResult> write, TaskCompletionSource<T> tcs, Message message, ServerEndPoint server)
+            {
+                var result = await write.ForAwait();
+                if (result != WriteResult.Success)
+                {
+                    var ex = @this.GetException(result, message, server);
+                    ThrowFailed(tcs, ex);
+                }
+                return tcs == null ? default : await tcs.Task.ForAwait();
+            }
+
+            TaskCompletionSource<T> tcs = null;
+            IResultBox<T> source = null;
+            if (!message.IsFireAndForget)
+            {
+                source = TaskResultBox<T>.Create(out tcs, state);
+            }
+            var write = TryPushMessageToBridgeAsync(message, processor, source, ref server);
+            if (!write.IsCompletedSuccessfully) return ExecuteAsyncImpl_Awaited(this, write, tcs, message, server);
+
+            if (tcs == null)
+            {
+                return CompletedTask<T>.Default(null); // F+F explicitly does not get async-state
+            }
+            else
+            {
+                var result = write.Result;
+                if (result != WriteResult.Success)
+                {
+                    var ex = GetException(result, message, server);
+                    ThrowFailed(tcs, ex);
+                }
+                return tcs.Task;
+            }
+        }
 
         internal void OnAsyncTimeout() => Interlocked.Increment(ref asyncTimeouts);
 
@@ -2786,9 +1908,11 @@ namespace StackExchange.Redis
         /// <returns>The number of instances known to have received the message (however, the actual number can be higher; returns -1 if the operation is pending).</returns>
         public long PublishReconfigure(CommandFlags flags = CommandFlags.None)
         {
-            byte[] channel = ConfigurationChangedChannel;
-            if (channel == null) return 0;
-            if (ReconfigureIfNeeded(null, false, "PublishReconfigure", true, flags))
+            if (ConfigurationChangedChannel is null)
+            {
+                return 0;
+            }
+            else if (ReconfigureIfNeeded(null, false, "PublishReconfigure", true, flags))
             {
                 return -1;
             }
@@ -2798,39 +1922,110 @@ namespace StackExchange.Redis
             }
         }
 
-        private long PublishReconfigureImpl(CommandFlags flags)
-        {
-            byte[] channel = ConfigurationChangedChannel;
-            if (channel == null) return 0;
-            return GetSubscriber().Publish(channel, RedisLiterals.Wildcard, flags);
-        }
+        private long PublishReconfigureImpl(CommandFlags flags) =>
+            ConfigurationChangedChannel is byte[] channel
+                ? GetSubscriber().Publish(channel, RedisLiterals.Wildcard, flags)
+                : 0;
 
         /// <summary>
         /// Sends request to all compatible clients to reconfigure or reconnect.
         /// </summary>
         /// <param name="flags">The command flags to use.</param>
         /// <returns>The number of instances known to have received the message (however, the actual number can be higher).</returns>
-        public Task<long> PublishReconfigureAsync(CommandFlags flags = CommandFlags.None)
-        {
-            byte[] channel = ConfigurationChangedChannel;
-            if (channel == null) return CompletedTask<long>.Default(null);
+        public Task<long> PublishReconfigureAsync(CommandFlags flags = CommandFlags.None) =>
+            ConfigurationChangedChannel is byte[] channel
+                ? GetSubscriber().PublishAsync(channel, RedisLiterals.Wildcard, flags)
+                : CompletedTask<long>.Default(null);
 
-            return GetSubscriber().PublishAsync(channel, RedisLiterals.Wildcard, flags);
+        /// <summary>
+        /// Release all resources associated with this object.
+        /// </summary>
+        public void Dispose()
+        {
+            GC.SuppressFinalize(this);
+            Close(!_isDisposed);
+            sentinelConnection?.Dispose();
+            var oldTimer = Interlocked.Exchange(ref sentinelPrimaryReconnectTimer, null);
+            oldTimer?.Dispose();
         }
 
         /// <summary>
-        /// Get the hash-slot associated with a given key, if applicable.
-        /// This can be useful for grouping operations.
+        /// Close all connections and release all resources associated with this object.
         /// </summary>
-        /// <param name="key">The <see cref="RedisKey"/> to determine the hash slot for.</param>
-        public int GetHashSlot(RedisKey key) => ServerSelectionStrategy.HashSlot(key);
-    }
+        /// <param name="allowCommandsToComplete">Whether to allow all in-queue commands to complete first.</param>
+        public void Close(bool allowCommandsToComplete = true)
+        {
+            if (_isDisposed) return;
 
-    internal enum WriteResult
-    {
-        Success,
-        NoConnectionAvailable,
-        TimeoutBeforeWrite,
-        WriteFailure,
+            OnClosing(false);
+            _isDisposed = true;
+            _profilingSessionProvider = null;
+            using (var tmp = pulse)
+            {
+                pulse = null;
+            }
+
+            if (allowCommandsToComplete)
+            {
+                var quits = QuitAllServers();
+                WaitAllIgnoreErrors(quits);
+            }
+            DisposeAndClearServers();
+            OnCloseReaderWriter();
+            OnClosing(true);
+            Interlocked.Increment(ref _connectionCloseCount);
+        }
+
+        /// <summary>
+        /// Close all connections and release all resources associated with this object.
+        /// </summary>
+        /// <param name="allowCommandsToComplete">Whether to allow all in-queue commands to complete first.</param>
+        public async Task CloseAsync(bool allowCommandsToComplete = true)
+        {
+            _isDisposed = true;
+            using (var tmp = pulse)
+            {
+                pulse = null;
+            }
+
+            if (allowCommandsToComplete)
+            {
+                var quits = QuitAllServers();
+                await WaitAllIgnoreErrorsAsync("quit", quits, RawConfig.AsyncTimeout, null).ForAwait();
+            }
+
+            DisposeAndClearServers();
+        }
+
+        private void DisposeAndClearServers()
+        {
+            lock (servers)
+            {
+                var iter = servers.GetEnumerator();
+                while (iter.MoveNext())
+                {
+                    var server = (ServerEndPoint)iter.Value;
+                    server.Dispose();
+                }
+                servers.Clear();
+            }
+        }
+
+        private Task[] QuitAllServers()
+        {
+            var quits = new Task[2 * servers.Count];
+            lock (servers)
+            {
+                var iter = servers.GetEnumerator();
+                int index = 0;
+                while (iter.MoveNext())
+                {
+                    var server = (ServerEndPoint)iter.Value;
+                    quits[index++] = server.Close(ConnectionType.Interactive);
+                    quits[index++] = server.Close(ConnectionType.Subscription);
+                }
+            }
+            return quits;
+        }
     }
 }
