+++ src/StackExchange.Redis/WriteResult.cs
@@ -1,0 +1,10 @@
+﻿namespace StackExchange.Redis
+{
+    internal enum WriteResult
+    {
+        Success,
+        NoConnectionAvailable,
+        TimeoutBeforeWrite,
+        WriteFailure,
+    }
+}
